_sffilbuf: squash another -Wsign-compare warning
The variable n is known to be greater than 0 in this branch.

+++ lib/sfio/sffilbuf.c
@@ -70,7 +70,7 @@ int _sffilbuf(Sfio_t * f, int n)
 	    if (n > 0) {
 		if (r > n && f->extent < 0 && (f->flags & SF_SHARE))
 		    r = n;	/* read only as much as requested */
-		else if (justseek && n <= f->iosz && f->iosz <= f->size)
+		else if (justseek && (size_t)n <= f->iosz && f->iosz <= f->size)
 		    r = f->iosz;	/* limit buffer filling */
 	    }
 	}
